feat: make PCST general and steiner tree a particular case
