+++ corneto/methods/future/pcst.py
@@ -1,26 +1,25 @@
-from typing import Any, List, Literal, Optional, Tuple, Union
+from typing import Any, Dict, List, Literal, Optional, Tuple, Union
 
 import numpy as np
 
 from corneto._constants import VAR_FLOW
 from corneto._data import Data
-from corneto._graph import BaseGraph, EdgeType
+from corneto._graph import Attr, BaseGraph, EdgeType
 from corneto.backend._base import Backend, ProblemDef
-from corneto.methods.future.steiner import SteinerTreeFlow
+from corneto.methods.future.method import FlowMethod
 
 
-class PrizeCollectingSteinerTree(SteinerTreeFlow):
+class PrizeCollectingSteinerTree(FlowMethod):
     """Prize-Collecting Steiner Tree optimization method as a flow-based problem.
 
-    This class extends the basic Steiner Tree algorithm with prize-collecting functionality.
     In a prize-collecting Steiner tree problem, terminals can have prizes (values > 0),
     making them optional terminals that provide a benefit if included in the solution.
-    The class now supports providing multiple max_flow and root_vertex values per sample.
+    The class supports providing multiple max_flow and root_vertex values per sample.
     """
 
     def __init__(
         self,
-        include_all_terminals: bool = True,
+        include_all_terminals: bool = False,
         max_flow: Optional[Union[float, List[float]]] = None,
         default_edge_cost: float = 1.0,
         flow_name: str = VAR_FLOW,
@@ -36,113 +35,290 @@ class PrizeCollectingSteinerTree(SteinerTreeFlow):
         backend: Optional[Backend] = None,
     ):
         super().__init__(
-            max_flow=max_flow,
-            default_edge_cost=default_edge_cost,
-            flow_name=flow_name,
-            root_vertex=root_vertex,
-            root_selection_strategy=root_selection_strategy,
-            epsilon=epsilon,
-            strict_acyclic=strict_acyclic,
-            disable_structured_sparsity=disable_structured_sparsity,
-            in_flow_edge_type=in_flow_edge_type,
-            out_flow_edge_type=out_flow_edge_type,
-            force_flow_through_root=force_flow_through_root,
             lambda_reg=lambda_reg,
+            disable_structured_sparsity=disable_structured_sparsity,
             backend=backend,
         )
         self.include_all_terminals = include_all_terminals
+        self.max_flow = max_flow
+        self.root_vertex = root_vertex
+        self.default_edge_cost = default_edge_cost
+        self.flow_name = flow_name
+        self.epsilon = epsilon
+        self.strict_acyclic = strict_acyclic
+        self.in_flow_edge_type = in_flow_edge_type
+        self.out_flow_edge_type = out_flow_edge_type
+        self.root_selection_strategy = root_selection_strategy
+        self.force_flow_through_root = force_flow_through_root
+
+        # Initialize containers and placeholders
+        self._terminal_edgeflow_idx = []
+        self.flow_edges = dict()
         self.prized_flow_edges = dict()
 
+        # Internal storage for per-sample values.
+        self._max_flow: Optional[float] = None
+        self._max_flow_list: Optional[List[float]] = None
+        self._root_vertex: Optional[Any] = None
+        self._root_vertex_list: Optional[List[Any]] = None
+
+        if isinstance(max_flow, list):
+            self._max_flow_list = max_flow
+        else:
+            self._max_flow = max_flow
+
+        if isinstance(root_vertex, list):
+            self._root_vertex_list = root_vertex
+        else:
+            self._root_vertex = root_vertex
+
+        self._selected_roots: List[Any] = []
+
     def preprocess(self, graph: BaseGraph, data: Data) -> Tuple[BaseGraph, Data]:
-        """Preprocess the graph and data for the Prize-Collecting Steiner tree optimization.
+        """Preprocess the graph and data."""
+        # Reset per-run attributes
+        self._terminal_edgeflow_idx = []
+        self.flow_edges = dict()
+        self.prized_flow_edges = dict()
 
-        Extends the base class preprocessing to handle prized nodes.
-        For each sample, we retrieve the sample-specific selected root from the base class
-        and add an out-edge for a prized node only if it is not that sample's root and
-        it does not already have a flow edge.
-        """
-        # Process the graph normally (this sets up terminals and sample-specific roots)
-        flow_graph, processed_data = super().preprocess(graph, data)
+        flow_graph = graph.copy()
+        all_vertices = data.query.filter_features(
+            lambda f: f.mapping == "vertex",
+        ).pluck_features()
+        terminals = data.query.filter_features(
+            lambda f: f.mapping == "vertex" and not f.value,
+        ).pluck_features()
 
-        # For each sample, add extra edges for prized nodes
-        for i, sample_data in enumerate(data.samples.values()):
-            sample_root = self._selected_roots[i]
-            prized_nodes = sample_data.query.select(lambda f: f.mapping == "vertex" and f.value).pluck()
-            for prized in prized_nodes:
-                # Only add an edge if the prized node is not the sample's root
-                # and it has not been already assigned a flow edge
-                if prized != sample_root and prized not in self.flow_edges:
-                    # TODO: If node not in the graph, raise an error, otherwise
-                    # this creates an infeasible problem.
-                    idx = flow_graph.add_edge(prized, (), type=self.out_flow_edge_type)
-                    self.flow_edges[prized] = idx
-                    self.prized_flow_edges[prized] = idx
+        num_samples = len(data.samples)
+        # Set up max_flow values per sample
+        if self._max_flow_list is not None:
+            if len(self._max_flow_list) != num_samples:
+                raise ValueError("Length of max_flow list must equal number of samples")
+        else:
+            if self._max_flow is None:
+                self._max_flow = len(all_vertices)
 
-        return flow_graph, processed_data
+        # Determine the root vertex to use for each sample.
+        selected_roots: List[Any] = []
+        if self._root_vertex_list is not None:
+            if len(self._root_vertex_list) != num_samples:
+                raise ValueError("Length of root_vertex list must equal number of samples")
+            for i in range(num_samples):
+                rv = self._root_vertex_list[i]
+                if rv is None:
+                    if self.root_selection_strategy == "first":
+                        chosen = next(iter(terminals)) if terminals else next(iter(all_vertices))
+                    elif self.root_selection_strategy == "best":
+                        chosen = None
+                    else:
+                        raise ValueError(f"Unknown root selection strategy: {self.root_selection_strategy}")
+                else:
+                    chosen = rv
+                selected_roots.append(chosen)
+        else:
+            if self._root_vertex is None:
+                if self.root_selection_strategy == "first":
+                    chosen = next(iter(terminals)) if terminals else next(iter(all_vertices))
+                elif self.root_selection_strategy == "best":
+                    chosen = None
+                else:
+                    raise ValueError(f"Unknown root selection strategy: {self.root_selection_strategy}")
+                selected_roots = [chosen] * num_samples
+            else:
+                selected_roots = [self._root_vertex] * num_samples
+
+        self._selected_roots = selected_roots
+
+        # Determine edge types based on the selected roots.
+        in_type = self.in_flow_edge_type
+        out_type = self.out_flow_edge_type
+        if all(r is None for r in selected_roots):
+            out_type = EdgeType.UNDIRECTED
+
+        # Create an in-edge for every unique root that is not None.
+        unique_roots = set(r for r in selected_roots if r is not None)
+        for r in unique_roots:
+            idx_root = flow_graph.add_edge((), r, type=in_type)
+            self.flow_edges[r] = idx_root
+
+        # For all other vertices, add an out-edge.
+        for v in all_vertices:
+            if v not in unique_roots:
+                idx = flow_graph.add_edge(v, (), type=out_type)
+                self.flow_edges[v] = idx
+
+        return flow_graph, data
+
+    def get_flow_bounds(self, graph: BaseGraph, data: Data) -> Dict[str, Any]:
+        """Get the flow bounds for the optimization problem."""
+        if self._max_flow_list is not None:
+            lb = [
+                np.array(
+                    [0 if prop.has_attr(Attr.EDGE_TYPE, EdgeType.DIRECTED) else -mf for prop in graph.get_attr_edges()]
+                )
+                for mf in self._max_flow_list
+            ]
+            ub = self._max_flow_list
+        else:
+            lb = np.array(
+                [
+                    0 if prop.has_attr(Attr.EDGE_TYPE, EdgeType.DIRECTED) else -self._max_flow
+                    for prop in graph.get_attr_edges()
+                ]
+            )
+            ub = self._max_flow
+
+        return {
+            "lb": lb,
+            "ub": ub,
+            "n_flows": len(data.samples),
+            "shared_bounds": False,
+        }
 
     def create_flow_based_problem(self, flow_problem: ProblemDef, graph: BaseGraph, data: Data):
-        """Create the flow-based Prize-Collecting Steiner tree optimization problem.
+        """Create the flow-based optimization problem."""
+        flow_edge_ids = list(self.flow_edges.values())
+        edge_ids = list(set(range(graph.num_edges)) - set(flow_edge_ids))
 
-        Builds on the base Steiner tree problem and adds prize-collecting functionality.
-        For each sample, after setting up the base flow problem (with sample-specific max_flow
-        and selected roots), we add objective terms to maximize the collected prizes.
-        """
-        # Create base Steiner tree problem (this uses per-sample max_flow and root values)
-        flow_problem = super().create_flow_based_problem(flow_problem, graph, data)
+        if self.strict_acyclic:
+            flow_problem += self.backend.NonZeroIndicator(flow_problem.expr._flow, tolerance=self.epsilon)
+            flow_problem += self.backend.Acyclic(
+                graph,
+                flow_problem,
+                indicator_negative_var_name="_flow_ineg",
+                indicator_positive_var_name="_flow_ipos",
+            )
+            with_flow = flow_problem.expr._flow_ipos + flow_problem.expr._flow_ineg
+        else:
+            flow_problem += self.backend.Indicator(flow_problem.expr._flow, indexes=edge_ids)
+            with_flow = flow_problem.expr._flow_i
 
-        # Add prize-collecting functionality for each sample.
+        flow_problem.register("with_flow", with_flow)
+        self._reg_varname = "with_flow"
+
         for i, sample_data in enumerate(data.samples.values()):
+            sample_max_flow = self._max_flow_list[i] if self._max_flow_list is not None else self._max_flow
+            sample_selected_root = self._selected_roots[i]
+
             F = flow_problem.expr.flow
             F = F if len(F.shape) == 1 else F[:, i]
 
-            # Retrieve prized terminals from the sample.
+            vertices_edgeflow_idx = []
+            all_vertices_with_data = sample_data.query.select(lambda f: f.mapping == "vertex").pluck()
+
+            terminals_edgeflow_idx = []
+            terminals = sample_data.query.select(
+                lambda f: f.mapping == "vertex" and f.data.get("role", None) == "terminal"
+            ).pluck()
+
             prized_terminals = dict(
                 sample_data.query.select(lambda f: f.mapping == "vertex" and f.value).pluck(lambda f: (f.id, f.value))
             )
 
+            for terminal in terminals:
+                if sample_selected_root is None or terminal != sample_selected_root:
+                    if terminal in self.flow_edges:
+                        idx = self.flow_edges[terminal]
+                        terminals_edgeflow_idx.append(idx)
+
+            for vertex in all_vertices_with_data:
+                if sample_selected_root is None or vertex != sample_selected_root:
+                    if vertex in self.flow_edges:
+                        idx = self.flow_edges[vertex]
+                        vertices_edgeflow_idx.append(idx)
+
+            sample_flow_edges = set(vertices_edgeflow_idx)
+            if sample_selected_root is not None:
+                sample_flow_edges.add(self.flow_edges[sample_selected_root])
+            sample_flow_edges = list(sample_flow_edges)
+            self._terminal_edgeflow_idx.append(sample_flow_edges)
+
+            other_flow_edges = list(set(self.flow_edges.values()) - set(sample_flow_edges))
+            if other_flow_edges:
+                flow_problem += F[other_flow_edges] == 0
+
+            # --- Root Flow Constraints ---
+            if sample_selected_root is not None:
+                if self.force_flow_through_root:
+                    flow_problem += F[self.flow_edges[sample_selected_root]] == sample_max_flow
+                else:
+                    flow_problem += F[self.flow_edges[sample_selected_root]] >= 0
+
+                if self.include_all_terminals and terminals_edgeflow_idx:
+                    flow_problem += F[terminals_edgeflow_idx] >= 1
+            else:
+                if all_vertices_with_data:
+                    flow_problem += self.backend.NonZeroIndicator(
+                        flow_problem.expr.flow,
+                        vertices_edgeflow_idx,
+                        i,
+                        tolerance=self.epsilon,
+                        suffix_pos=f"_terminal_pos_{i}",
+                        suffix_neg=f"_terminal_neg_{i}",
+                    )
+                    terminal_pos = self.flow_name + f"_terminal_pos_{i}"
+                    terminal_neg = self.flow_name + f"_terminal_neg_{i}"
+
+                    flow_problem += flow_problem.expr[terminal_neg].sum() == 1
+
+                    if self.include_all_terminals:
+                        t_idx = [vertices_edgeflow_idx.index(idx) for idx in terminals_edgeflow_idx]
+                        if t_idx:
+                            flow_problem += flow_problem.expr[terminal_pos][t_idx].sum() == len(t_idx) - 1
+
+            # --- Costs Objective ---
+            edge_costs = np.ones((len(edge_ids))) * self.default_edge_cost
+            selected = with_flow if len(with_flow.shape) == 1 else with_flow[:, i]
+
+            edge_data = sample_data.query.select(lambda f: f.mapping == "edge").to_list()
+            for edata in edge_data:
+                edge_costs[edata.id] = float(edata.value)
+
+            flow_problem.add_objective(edge_costs[edge_ids] @ selected[edge_ids], name="edge_cost")
+
+            # --- Prizes Objective (PCST) ---
             if prized_terminals:
                 prized_idx = [
                     self.flow_edges[prized] for prized in prized_terminals.keys() if prized in self.flow_edges
                 ]
                 if prized_idx:
-                    # Build a prize vector (order must match the prized_idx)
                     prizes = np.array(
                         [prized_terminals[prized] for prized in prized_terminals.keys() if prized in self.flow_edges]
                     )
+
                     if self.strict_acyclic:
-                        selected = flow_problem.expr._flow_ipos + flow_problem.expr._flow_ineg
-                        selected = selected if len(selected.shape) == 1 else selected[:, i]
-                        selected_prized_flow_edges = selected[prized_idx]
+                        selected_for_prizes = flow_problem.expr._flow_ipos + flow_problem.expr._flow_ineg
+                        selected_for_prizes = (
+                            selected_for_prizes if len(selected_for_prizes.shape) == 1 else selected_for_prizes[:, i]
+                        )
+                        selected_prized_flow_edges = selected_for_prizes[prized_idx]
                     else:
                         indicator_terminal_pos = self.flow_name + f"_terminal_pos_{i}"
                         indicator_terminal_neg = self.flow_name + f"_terminal_neg_{i}"
-                        if (
-                            indicator_terminal_pos not in flow_problem.expr
-                            or indicator_terminal_neg not in flow_problem.expr
-                        ):
+
+                        if sample_selected_root is None:
+                            flow_edges_idxs = vertices_edgeflow_idx
+                            indices_in_indicator = [flow_edges_idxs.index(idx) for idx in prized_idx]
+                            selected_vec = (
+                                flow_problem.expr[indicator_terminal_pos] + flow_problem.expr[indicator_terminal_neg]
+                            )
+                            selected_vec = selected_vec if len(selected_vec.shape) == 1 else selected_vec[:, i]
+                            selected_prized_flow_edges = selected_vec[indices_in_indicator]
+                        else:
                             flow_problem += self.backend.NonZeroIndicator(
-                                flow_problem.expr.flow,
+                                F,
                                 prized_idx,
                                 i,
                                 tolerance=self.epsilon,
-                                suffix_pos=f"_terminal_pos_{i}",
-                                suffix_neg=f"_terminal_neg_{i}",
+                                suffix_pos=f"_prize_pos_{i}",
+                                suffix_neg=f"_prize_neg_{i}",
                             )
                             selected_prized_flow_edges = (
-                                flow_problem.expr[indicator_terminal_pos] + flow_problem.expr[indicator_terminal_neg]
+                                flow_problem.expr[f"{self.flow_name}_prize_pos_{i}"]
+                                + flow_problem.expr[f"{self.flow_name}_prize_neg_{i}"]
                             )
-                        else:
-                            flow_edges_idxs = self._terminal_edgeflow_idx[i]
-                            prized_idx_in_list = [flow_edges_idxs.index(idx) for idx in prized_idx]
-                            selected = (
-                                flow_problem.expr[indicator_terminal_pos] + flow_problem.expr[indicator_terminal_neg]
-                            )
-                            selected = selected if len(selected.shape) == 1 else selected[:, i]
-                            selected_prized_flow_edges = selected[prized_idx_in_list]
 
                     flow_problem.register(f"selected_prized_flow_edges_{i}", selected_prized_flow_edges)
-
-                    # Add an objective term (with negative weight) to maximize prizes.
                     flow_problem.add_objective(prizes @ selected_prized_flow_edges, weight=-1, name="prizes")
 
         return flow_problem
